Added CopyModelStep interface, added clustering keeping parcel bundles in clusters
implement CopyModelStep interface for all model steps that can support it:
- ActiveDeliveryPartnerSelector
- DistributionCenterSelectorByFleetSize
- MultipleModelOptionsStep
- RandomDateSelector
- ShareBasedSelector
- VolumeSelector
These model steps now allow to set a "copyGetter" in the ParcelDemandModelBuilder.
(Other (non-copy-bundle) model steps throw "UnsupportedOperation")
ParcelDemandModelStep now checks if copyGetter is set, and if the list of previous bundle parcels is non-empty: if that is the case, the copy getter is applied to the first bundle parcel
LatentModelStepWarpper delegates bundleCopy setter to the wrapped step.

steps that might require CopyModelStep in the future:
- CopyProviderModelStep

Added KeepBundlesClustering as a wrapper for DeliveryClusteringStrategy.
It defines that two parcels can be grouped, if their bundle id is equal, or if the delegate clustering strategy considers them groupable.
Also, KeepBundlesClustering overrides the cluster method, making sure that parcels of the same bundle end up in the same partitioned cluster. Thi might violate the max cluster size, if a bundle with more parcels exists.

Added QuantityAndBundlingCsvReader to parse csv files containing both quantity and bundling as columns.
The bundling column is expected to contain a ";" separated List of integers, each denoting the size of one bundle. Sum of bundle sizes should match the value in the quantity column.
The list of bundle sizes must be a quoted string in the csv, since the csv separator is also ";".

renamed "shipmentSize" methods in ParcelDemandModelBuilder to "parcelSize"